--- src/main/kotlin/com/example/gsmkotlin/GsmKotlinApplication.kt
@@ -1,11 +1,0 @@
-package com.example.gsmkotlin
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class GsmKotlinApplication
-
-fun main(args: Array<String>) {
-    runApplication<GsmKotlinApplication>(*args)
-}
