--- src/main/kotlin/com/example/gsmkotlin/domain/user/type/Authority.kt
@@ -1,5 +1,0 @@
-package com.example.gsmkotlin.domain.user.type
-
-enum class Authority {
-    STUDENT, TEACHER, ADMIN
-}
