--- src/main/kotlin/com/example/gsmkotlin/global/error/ErrorResponse.kt
@@ -1,11 +1,0 @@
-package com.example.gsmkotlin.global.error
-
-class ErrorResponse (
-    val message: String
-) {
-    companion object {
-        fun of(cause: Throwable): ErrorResponse {
-            return ErrorResponse(cause.message ?: "An unknown error occurred")
-        }
-    }
-}
