--- src/main/kotlin/com/example/gsmkotlin/global/error/GlobalException.kt
@@ -1,10 +1,0 @@
-package com.example.gsmkotlin.global.error
-
-import org.springframework.http.HttpStatus
-
-class GlobalException (
-    override val message: String,
-    val httpStatus: HttpStatus
-) : RuntimeException(message) {
-    constructor(httpStatus: HttpStatus) : this(httpStatus.reasonPhrase, httpStatus)
-}
