--- src/main/kotlin/com/example/gsmkotlin/global/security/auth/CustomUserDetails.kt
@@ -1,19 +1,0 @@
-package com.example.gsmkotlin.global.security.auth
-
-import com.example.gsmkotlin.domain.user.entity.User
-import org.springframework.security.core.GrantedAuthority
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.userdetails.UserDetails
-
-class CustomUserDetails (
-    private val user: User
-) : UserDetails {
-    override fun getAuthorities(): MutableCollection<out GrantedAuthority> =
-        mutableListOf(SimpleGrantedAuthority(user.authority.name))
-    override fun getPassword(): String? = null
-    override fun getUsername(): String = user.name
-    override fun isAccountNonExpired(): Boolean = true
-    override fun isAccountNonLocked(): Boolean = true
-    override fun isCredentialsNonExpired(): Boolean = true
-    override fun isEnabled(): Boolean = true
-}
