--- src/main/kotlin/com/example/gsmkotlin/global/security/config/SecurityConfig.kt
@@ -1,69 +1,0 @@
-package com.example.gsmkotlin.global.security.config
-
-import com.example.gsmkotlin.global.security.handler.CustomAccessDeniedHandler
-import com.example.gsmkotlin.global.security.handler.CustomAuthenticationEntryPointHandler
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.http.HttpMethod
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.web.SecurityFilterChain
-import org.springframework.web.cors.CorsConfiguration
-import org.springframework.web.cors.CorsConfigurationSource
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource
-import java.util.*
-
-@Configuration
-class SecurityConfig (
-    private val accessDeniedHandler: CustomAccessDeniedHandler,
-    private val authenticationEntryPointHandler: CustomAuthenticationEntryPointHandler
-) {
-    
-    @Bean
-    @Throws(Exception::class)
-    fun filterChain(http: HttpSecurity): SecurityFilterChain {
-        
-        http.formLogin { formLogin -> formLogin.disable() }
-            .httpBasic { httpBasic -> httpBasic.disable() }
-        
-        http.csrf { csrf -> csrf.disable() }
-            .cors { cors -> cors.configurationSource(corsConfigurationSource()) }
-        
-        http.exceptionHandling { handling -> handling
-            .accessDeniedHandler(accessDeniedHandler)
-            .authenticationEntryPoint(authenticationEntryPointHandler)
-        }
-        
-        http.sessionManagement { sessionManagement -> sessionManagement
-            .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        }
-        
-        http.authorizeHttpRequests { httpRequests -> httpRequests
-            .anyRequest().permitAll()
-        }
-        
-        return http.build()
-    }
-    
-    @Bean
-    fun corsConfigurationSource(): CorsConfigurationSource {
-        val configuration = CorsConfiguration()
-        
-        // plz custom allowed client origins
-        configuration.allowedOrigins = listOf("*")
-        configuration.setAllowedMethods(
-            listOf(
-                HttpMethod.GET.name(),
-                HttpMethod.POST.name(),
-                HttpMethod.PUT.name(),
-                HttpMethod.PATCH.name(),
-                HttpMethod.DELETE.name(),
-                HttpMethod.OPTIONS.name()
-            )
-        )
-        configuration.allowCredentials = true
-        val source = UrlBasedCorsConfigurationSource()
-        source.registerCorsConfiguration("/**",configuration)
-        return source
-    }
-}
