+++ src/main/kotlin/gsm/gsmkotlin/GsmKotlinApplication.kt
@@ -1,0 +1,11 @@
+package gsm.gsmkotlin
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class GsmKotlinApplication
+
+fun main(args: Array<String>) {
+    runApplication<GsmKotlinApplication>(*args)
+}
