+++ src/main/kotlin/gsm/gsmkotlin/domain/user/repository/UserRepository.kt
@@ -1,0 +1,8 @@
+package gsm.gsmkotlin.domain.user.repository
+
+import gsm.gsmkotlin.domain.user.entity.User
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository: JpaRepository<User, Long> {
+    fun findByEmail(email: String): User?
+}
