+++ src/main/kotlin/gsm/gsmkotlin/domain/user/type/Authority.kt
@@ -1,0 +1,5 @@
+package gsm.gsmkotlin.domain.user.type
+
+enum class Authority {
+    STUDENT, TEACHER, ADMIN
+}
