+++ src/main/kotlin/gsm/gsmkotlin/global/config/PropertiesScanConfig.kt
@@ -1,0 +1,13 @@
+package gsm.gsmkotlin.global.config
+
+import gsm.gsmkotlin.global.security.jwt.properties.JwtEnvironment
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@ConfigurationPropertiesScan(
+    basePackageClasses = [
+        JwtEnvironment::class
+    ]
+)
+class PropertiesScanConfig
