+++ src/main/kotlin/gsm/gsmkotlin/global/error/GlobalException.kt
@@ -1,0 +1,10 @@
+package gsm.gsmkotlin.global.error
+
+import org.springframework.http.HttpStatus
+
+class GlobalException (
+    override val message: String,
+    val httpStatus: HttpStatus
+) : RuntimeException(message) {
+    constructor(httpStatus: HttpStatus) : this(httpStatus.reasonPhrase, httpStatus)
+}
