+++ src/main/kotlin/gsm/gsmkotlin/global/security/handler/CustomAccessDeniedHandler.kt
@@ -1,0 +1,29 @@
+package gsm.gsmkotlin.global.security.handler
+
+import gsm.gsmkotlin.global.error.ErrorResponse
+import com.fasterxml.jackson.databind.ObjectMapper
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.slf4j.LoggerFactory
+import org.springframework.http.HttpStatus
+import org.springframework.http.MediaType
+import org.springframework.security.access.AccessDeniedException
+import org.springframework.security.web.access.AccessDeniedHandler
+import org.springframework.stereotype.Component
+
+@Component
+class CustomAccessDeniedHandler : AccessDeniedHandler {
+    
+    private val log = LoggerFactory.getLogger(this::class.simpleName)
+    
+    override fun handle(request: HttpServletRequest, response: HttpServletResponse, accessDeniedException: AccessDeniedException) {
+        log.info("========== ACCESS DENIED ==========")
+        
+        response.characterEncoding = "utf-8"
+        response.status = HttpStatus.FORBIDDEN.value()
+        response.contentType = MediaType.APPLICATION_JSON_VALUE
+        response.writer.write(ObjectMapper().writeValueAsString(
+            ErrorResponse("요청을 수행할 수 있는 권한이 없습니다.")
+        ))
+    }
+}
