+++ src/main/kotlin/gsm/gsmkotlin/global/security/jwt/TokenParser.kt
@@ -1,0 +1,32 @@
+package gsm.gsmkotlin.global.security.jwt
+
+import gsm.gsmkotlin.global.security.auth.CustomUserDetailsService
+import gsm.gsmkotlin.global.security.jwt.properties.JwtEnvironment
+import io.jsonwebtoken.Claims
+import io.jsonwebtoken.Jwts
+import io.jsonwebtoken.security.Keys
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
+import org.springframework.stereotype.Component
+import java.nio.charset.StandardCharsets
+import java.security.Key
+
+@Component
+class TokenParser (
+    private val userDetailsService: CustomUserDetailsService,
+    private val jwtEnv: JwtEnvironment
+) {
+    fun authentication(accessToken: String): UsernamePasswordAuthenticationToken {
+        val authenticationDetails = userDetailsService.loadUserByUsername(getAccessTokenSubject(accessToken))
+        return UsernamePasswordAuthenticationToken(authenticationDetails, "", authenticationDetails.authorities)
+    }
+    
+    private fun getAccessTokenSubject(accessToken: String): String =
+        getTokenBody(accessToken, Keys.hmacShaKeyFor(jwtEnv.accessSecret.toByteArray(StandardCharsets.UTF_8))).subject
+    
+    private fun getTokenBody(token: String, secret: Key): Claims =
+        Jwts.parserBuilder()
+            .setSigningKey(secret)
+            .build()
+            .parseClaimsJws(token)
+            .body
+}
