+++ src/main/kotlin/gsm/gsmkotlin/global/security/jwt/dto/TokenDto.kt
@@ -1,0 +1,8 @@
+package gsm.gsmkotlin.global.security.jwt.dto
+
+data class TokenDto (
+    val accessToken: String,
+    val refreshToken: String,
+    val accessTokenExp: Int,
+    val refreshTokenExp: Int
+)
