--- src/test/kotlin/com/example/gsmkotlin/GsmKotlinApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.gsmkotlin
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class GsmKotlinApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
