+++ src/test/kotlin/gsm/gsmkotlin/GsmKotlinApplicationTests.kt
@@ -1,0 +1,13 @@
+package gsm.gsmkotlin
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class GsmKotlinApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
